Faster R-CNN/requirements.txt

datasets

# Core packages
numpy
pandas
pathlib

# Deep Learning
tensorflow[gpu]  # This will handle both CPU and GPU versions

# Image processing
Pillow  # Required by ImageDataGenerator

# Machine Learning
scikit-learn

# Visualization
matplotlib
seaborn  # For statistical data visualization

# Utilities
tqdm  # For progress bars
typing-extensions  # For type hints
argparse

# Optional but recommended
ipython  # For interactive development
jupyter  # For notebook support

# Add to requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.11.0
pyyaml
torch 
torchvision 
tensorboard 
torchmetrics 

opencv-python
pycocotools